+++ backend/src/main/java/com/allog/dallog/domain/subscription/application/SubscriptionService.java
@@ -2,9 +2,9 @@ package com.allog.dallog.domain.subscription.application;
 
 import com.allog.dallog.domain.category.domain.Category;
 import com.allog.dallog.domain.category.domain.CategoryRepository;
-import com.allog.dallog.domain.categoryrole.CategoryRole;
-import com.allog.dallog.domain.categoryrole.CategoryRoleRepository;
-import com.allog.dallog.domain.categoryrole.CategoryRoleType;
+import com.allog.dallog.domain.categoryrole.domain.CategoryRole;
+import com.allog.dallog.domain.categoryrole.domain.CategoryRoleRepository;
+import com.allog.dallog.domain.categoryrole.domain.CategoryRoleType;
 import com.allog.dallog.domain.member.domain.Member;
 import com.allog.dallog.domain.member.domain.MemberRepository;
 import com.allog.dallog.domain.subscription.domain.Color;
