+++ backend/src/test/java/com/allog/dallog/domain/member/application/MemberServiceTest.java
@@ -13,9 +13,9 @@ import com.allog.dallog.common.fixtures.AuthFixtures;
 import com.allog.dallog.common.fixtures.SubscriptionFixtures;
 import com.allog.dallog.domain.category.domain.Category;
 import com.allog.dallog.domain.category.domain.CategoryRepository;
-import com.allog.dallog.domain.categoryrole.CategoryRole;
-import com.allog.dallog.domain.categoryrole.CategoryRoleRepository;
-import com.allog.dallog.domain.categoryrole.CategoryRoleType;
+import com.allog.dallog.domain.categoryrole.domain.CategoryRole;
+import com.allog.dallog.domain.categoryrole.domain.CategoryRoleRepository;
+import com.allog.dallog.domain.categoryrole.domain.CategoryRoleType;
 import com.allog.dallog.domain.member.domain.Member;
 import com.allog.dallog.domain.member.domain.MemberRepository;
 import com.allog.dallog.domain.member.dto.MemberResponse;
